remove spirv-builder-cli, leave install broken
