--- crates/spirv-builder-cli/Cargo.toml
@@ -1,61 +1,0 @@
-[package]
-name = "spirv-builder-cli"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-# `clap = "=4.4.8"` pin is needed to support older Rust toolchains.
-# If `clap` gets too old, then it can be split into 2 versions, gated by a feature, like `spirv` is.
-# Or more generally we should think about how to deprecate supporting old version of `rust-gpu`.
-clap = { version = "=4.4.8", features = ["derive"] }
-env_home = "0.1.0"
-env_logger = "0.10"
-log = "0.4"
-serde = "1.0.214"
-serde_json = "1.0.132"
-toml = "0.8.19"
-
-[features]
-default = ["spirv-builder-0_10"]
-# The `spirv-builder` before `cargo gpu` existed. It has an incompatible `SpirvBuilder` interface.
-spirv-builder-pre-cli = ["dep:spirv-builder-pre-cli", "dep:spirv_0_2"]
-# The first version that introduced `cargo gpu`. It has some extra `.builder()` args that make
-# dynamically changing build dependencies easier.
-spirv-builder-0_10 = ["dep:spirv-builder-0_10", "dep:spirv_0_3"]
-#
-rspirv-latest = ["dep:spirv_0_3"]
-
-[dependencies.spirv_0_2]
-package = "spirv"
-version = "0.2.0"
-features = [ "deserialize", "serialize" ]
-optional = true
-
-[dependencies.spirv_0_3]
-package = "spirv"
-version = "0.3.0"
-features = [ "deserialize", "serialize" ]
-optional = true
-
-# NB: All the `${AUTO-REPLACE*}` tokens in each feature get replaced with the same values.
-# This is because only one feature can ever be used at once and it makes it easier to just
-# replace each token rather than figure out to what feature each token belongs.
-
-[dependencies.spirv-builder-pre-cli]
-package = "spirv-builder"
-features = [ "watch" ]
-optional = true
-git = "https://github.com/Rust-GPU/rust-gpu" # ${AUTO-REPLACE-SOURCE}
-rev = "4c633aec" # ${AUTO-REPLACE-VERSION}
-
-[dependencies.spirv-builder-0_10]
-package = "spirv-builder"
-features = [ "watch" ]
-optional = true
-git = "https://github.com/Rust-GPU/rust-gpu" # ${AUTO-REPLACE-SOURCE}
-rev = "60dcb82" # ${AUTO-REPLACE-VERSION}
- 
-[lints.rust]
-# This crate is most often run by end users compiling their shaders so it's not so relevant
-# for them to see warnings.
-warnings = "allow"
