--- crates/spirv-builder-cli/README.md
@@ -1,7 +1,0 @@
-# SPIR-V Builder CLI Template
-
-> [!CAUTION]
-> This may look like a typical crate, but it is in fact more like a crate _template_.
-> It is intended to be primarily called by `cargo gpu`, which has these files embedded
-> into its binary and so can copy and edit them to automatically compile end-user
-> shaders.
