--- crates/spirv-builder-cli/rust-toolchain.toml
@@ -1,5 +1,0 @@
-# This is just a placeholder for development. In actual usage the toolchain
-# is set dynamically.
-[toolchain]
-channel = "nightly-2024-04-24"
-components = ["rust-src", "rustc-dev", "llvm-tools"]
